added light player code and tidying up folders
