+++ test/Surround_Test_Audio.ass
@@ -1,0 +1,57 @@
+﻿[Script Info]
+; Script generated by Aegisub 3.2.2
+; http://www.aegisub.org/
+Title: Default Aegisub file
+ScriptType: v4.00+
+WrapStyle: 0
+ScaledBorderAndShadow: yes
+YCbCr Matrix: TV.601
+PlayResX: 1920
+PlayResY: 1080
+
+[Aegisub Project Garbage]
+Last Style Storage: Default
+Audio File: Surround_Test_Video.mp4
+Video File: Surround_Test_Video.mp4
+Video AR Mode: 4
+Video AR Value: 1.777778
+Video Zoom Percent: 0.500000
+Scroll Position: 19
+Active Line: 28
+Video Position: 1034
+
+[V4+ Styles]
+Format: Name, Fontname, Fontsize, PrimaryColour, SecondaryColour, OutlineColour, BackColour, Bold, Italic, Underline, StrikeOut, ScaleX, ScaleY, Spacing, Angle, BorderStyle, Outline, Shadow, Alignment, MarginL, MarginR, MarginV, Encoding
+Style: Default,Arial,40,&H00FFFFFF,&H000000FF,&H00000000,&H00000000,0,0,0,0,100,100,0,0,1,2,2,2,10,10,10,1
+
+[Events]
+Format: Layer, Start, End, Style, Name, MarginL, MarginR, MarginV, Effect, Text
+Dialogue: 0,0:00:00.00,0:00:02.05,Default,,0,0,0,,HUE1(0,255,0,0);HUE2(0,255,0,0);HUE3(0,255,0,0);DMX1(0, 0, 0, 0)
+Dialogue: 0,0:00:02.19,0:00:03.02,Default,,0,0,0,,HUE1(5000,255,100,0);HUE2(5000,255,100,0);HUE3(5000,255,100,0);DMX1(255, 0, 0, 0)
+Dialogue: 0,0:00:03.19,0:00:04.02,Default,,0,0,0,,HUE1(10000,255,100,0);HUE2(10000,255,100,0);HUE3(10000,255,100,0);DMX1(0, 255, 0, 0)
+Dialogue: 0,0:00:04.21,0:00:05.01,Default,,0,0,0,,HUE1(15000,255,100,0);HUE2(15000,255,100,0);HUE3(15000,255,100,0);DMX1(0, 0, 255, 0)
+Dialogue: 0,0:00:05.23,0:00:06.19,Default,,0,0,0,,HUE1(20000,255,100,0);HUE2(20000,255,100,0);HUE3(20000,255,100,0);DMX1(0, 0, 0, 255)
+Dialogue: 0,0:00:06.29,0:00:07.20,Default,,0,0,0,,HUE1(25000,255,100,0);HUE2(25000,255,100,0);HUE3(25000,255,100,0);DMX1(255, 0, 255, 0)
+Dialogue: 0,0:00:07.24,0:00:09.76,Default,,0,0,0,,HUE1(30000,255,100,0);HUE2(30000,255,100,0);HUE3(30000,255,100,0);DMX1(255, 255, 255, 255)
+Dialogue: 0,0:00:10.06,0:00:10.46,Default,,0,0,0,,HUE1(0,255,0,0);HUE2(0,255,0,0);HUE3(0,255,0,0);DMX1(0, 0, 0, 0)
+Dialogue: 0,0:00:10.57,0:00:11.37,Default,,0,0,0,,HUE1(5000,255,100,0);HUE2(5000,255,100,0);HUE3(5000,255,100,0);DMX1(255, 0, 0, 0)
+Dialogue: 0,0:00:11.51,0:00:12.31,Default,,0,0,0,,HUE1(10000,255,100,0);HUE2(10000,255,100,0);HUE3(10000,255,100,0);DMX1(0, 255, 0, 0)
+Dialogue: 0,0:00:12.48,0:00:13.37,Default,,0,0,0,,HUE1(15000,255,100,0);HUE2(15000,255,100,0);HUE3(15000,255,100,0);DMX1(0, 0, 255, 0)
+Dialogue: 0,0:00:13.54,0:00:14.51,Default,,0,0,0,,HUE1(20000,255,100,0);HUE2(20000,255,100,0);HUE3(20000,255,100,0);DMX1(0, 0, 0, 255)
+Dialogue: 0,0:00:14.54,0:00:15.60,Default,,0,0,0,,HUE1(25000,255,100,0);HUE2(25000,255,100,0);HUE3(25000,255,100,0);DMX1(255, 0, 255, 0)
+Dialogue: 0,0:00:15.82,0:00:17.85,Default,,0,0,0,,HUE1(30000,255,100,0);HUE2(30000,255,100,0);HUE3(30000,255,100,0);DMX1(255, 255, 255, 255)
+Dialogue: 0,0:00:09.94,0:00:10.42,Default,,0,0,0,,HUE1(0,255,0,0);HUE2(0,255,0,0);HUE3(0,255,0,0);DMX1(0, 0, 0, 0)
+Dialogue: 0,0:00:18.75,0:00:19.55,Default,,0,0,0,,HUE1(5000,255,100,0);HUE2(5000,255,100,0);HUE3(5000,255,100,0);DMX1(255, 0, 0, 0)
+Dialogue: 0,0:00:19.77,0:00:20.60,Default,,0,0,0,,HUE1(10000,255,100,0);HUE2(10000,255,100,0);HUE3(10000,255,100,0);DMX1(0, 255, 0, 0)
+Dialogue: 0,0:00:20.77,0:00:21.57,Default,,0,0,0,,HUE1(15000,255,100,0);HUE2(15000,255,100,0);HUE3(15000,255,100,0);DMX1(0, 0, 255, 0)
+Dialogue: 0,0:00:21.85,0:00:22.80,Default,,0,0,0,,HUE1(20000,255,100,0);HUE2(20000,255,100,0);HUE3(20000,255,100,0);DMX1(0, 0, 0, 255)
+Dialogue: 0,0:00:22.85,0:00:23.91,Default,,0,0,0,,HUE1(25000,255,100,0);HUE2(25000,255,100,0);HUE3(25000,255,100,0);DMX1(255, 0, 255, 0)
+Dialogue: 0,0:00:23.94,0:00:26.22,Default,,0,0,0,,HUE1(30000,255,100,0);HUE2(30000,255,100,0);HUE3(30000,255,100,0);DMX1(255, 255, 255, 255)
+Dialogue: 0,0:00:26.42,0:00:27.08,Default,,0,0,0,,HUE1(0,255,0,0);HUE2(0,255,0,0);HUE3(0,255,0,0);DMX1(0, 0, 0, 0)
+Dialogue: 0,0:00:27.14,0:00:27.94,Default,,0,0,0,,HUE1(5000,255,100,0);HUE2(5000,255,100,0);HUE3(5000,255,100,0);DMX1(255, 0, 0, 0)
+Dialogue: 0,0:00:28.08,0:00:28.88,Default,,0,0,0,,HUE1(10000,255,100,0);HUE2(10000,255,100,0);HUE3(10000,255,100,0);DMX1(0, 255, 0, 0)
+Dialogue: 0,0:00:29.17,0:00:30.11,Default,,0,0,0,,HUE1(15000,255,100,0);HUE2(15000,255,100,0);HUE3(15000,255,100,0);DMX1(0, 0, 255, 0)
+Dialogue: 0,0:00:30.11,0:00:31.14,Default,,0,0,0,,HUE1(20000,255,100,0);HUE2(20000,255,100,0);HUE3(20000,255,100,0);DMX1(0, 0, 0, 255)
+Dialogue: 0,0:00:31.14,0:00:32.22,Default,,0,0,0,,HUE1(25000,255,100,0);HUE2(25000,255,100,0);HUE3(25000,255,100,0);DMX1(255, 0, 255, 0)
+Dialogue: 0,0:00:32.34,0:00:34.34,Default,,0,0,0,,HUE1(30000,255,100,0);HUE2(30000,255,100,0);HUE3(30000,255,100,0);DMX1(255, 255, 255, 255)
+Dialogue: 0,0:00:34.48,0:00:37.40,Default,,0,0,0,,HUE1(0,255,0,0);HUE2(0,255,0,0);HUE3(0,255,0,0);DMX1(0, 0, 0, 0)
