+++ test/Surround_Test_Audio.srt
@@ -1,0 +1,121 @@
+﻿1
+00:00:00,000 --> 00:00:02,050
+HUE1(0,255,0,0);HUE2(0,255,0,0);HUE3(0,255,0,0);DMX1(0, 0, 0, 0)
+
+2
+00:00:02,190 --> 00:00:03,020
+HUE1(5000,255,100,0);HUE2(5000,255,100,0);HUE3(5000,255,100,0);DMX1(255, 0, 0, 0)
+
+3
+00:00:03,190 --> 00:00:04,020
+HUE1(10000,255,100,0);HUE2(10000,255,100,0);HUE3(10000,255,100,0);DMX1(0, 255, 0, 0)
+
+4
+00:00:04,210 --> 00:00:05,010
+HUE1(15000,255,100,0);HUE2(15000,255,100,0);HUE3(15000,255,100,0);DMX1(0, 0, 255, 0)
+
+5
+00:00:05,230 --> 00:00:06,190
+HUE1(20000,255,100,0);HUE2(20000,255,100,0);HUE3(20000,255,100,0);DMX1(0, 0, 0, 255)
+
+6
+00:00:06,290 --> 00:00:07,200
+HUE1(25000,255,100,0);HUE2(25000,255,100,0);HUE3(25000,255,100,0);DMX1(255, 0, 255, 0)
+
+7
+00:00:07,240 --> 00:00:09,760
+HUE1(30000,255,100,0);HUE2(30000,255,100,0);HUE3(30000,255,100,0);DMX1(255, 255, 255, 255)
+
+8
+00:00:09,940 --> 00:00:10,060
+HUE1(0,255,0,0);HUE2(0,255,0,0);HUE3(0,255,0,0);DMX1(0, 0, 0, 0)
+
+9
+00:00:10,060 --> 00:00:10,420
+HUE1(0,255,0,0);HUE2(0,255,0,0);HUE3(0,255,0,0);DMX1(0, 0, 0, 0)
+HUE1(0,255,0,0);HUE2(0,255,0,0);HUE3(0,255,0,0);DMX1(0, 0, 0, 0)
+
+10
+00:00:10,420 --> 00:00:10,464
+HUE1(0,255,0,0);HUE2(0,255,0,0);HUE3(0,255,0,0);DMX1(0, 0, 0, 0)
+
+11
+00:00:10,571 --> 00:00:11,371
+HUE1(5000,255,100,0);HUE2(5000,255,100,0);HUE3(5000,255,100,0);DMX1(255, 0, 0, 0)
+
+12
+00:00:11,514 --> 00:00:12,314
+HUE1(10000,255,100,0);HUE2(10000,255,100,0);HUE3(10000,255,100,0);DMX1(0, 255, 0, 0)
+
+13
+00:00:12,485 --> 00:00:13,371
+HUE1(15000,255,100,0);HUE2(15000,255,100,0);HUE3(15000,255,100,0);DMX1(0, 0, 255, 0)
+
+14
+00:00:13,542 --> 00:00:14,514
+HUE1(20000,255,100,0);HUE2(20000,255,100,0);HUE3(20000,255,100,0);DMX1(0, 0, 0, 255)
+
+15
+00:00:14,542 --> 00:00:15,600
+HUE1(25000,255,100,0);HUE2(25000,255,100,0);HUE3(25000,255,100,0);DMX1(255, 0, 255, 0)
+
+16
+00:00:15,828 --> 00:00:17,857
+HUE1(30000,255,100,0);HUE2(30000,255,100,0);HUE3(30000,255,100,0);DMX1(255, 255, 255, 255)
+
+17
+00:00:18,750 --> 00:00:19,550
+HUE1(5000,255,100,0);HUE2(5000,255,100,0);HUE3(5000,255,100,0);DMX1(255, 0, 0, 0)
+
+18
+00:00:19,771 --> 00:00:20,600
+HUE1(10000,255,100,0);HUE2(10000,255,100,0);HUE3(10000,255,100,0);DMX1(0, 255, 0, 0)
+
+19
+00:00:20,771 --> 00:00:21,571
+HUE1(15000,255,100,0);HUE2(15000,255,100,0);HUE3(15000,255,100,0);DMX1(0, 0, 255, 0)
+
+20
+00:00:21,857 --> 00:00:22,800
+HUE1(20000,255,100,0);HUE2(20000,255,100,0);HUE3(20000,255,100,0);DMX1(0, 0, 0, 255)
+
+21
+00:00:22,857 --> 00:00:23,914
+HUE1(25000,255,100,0);HUE2(25000,255,100,0);HUE3(25000,255,100,0);DMX1(255, 0, 255, 0)
+
+22
+00:00:23,942 --> 00:00:26,228
+HUE1(30000,255,100,0);HUE2(30000,255,100,0);HUE3(30000,255,100,0);DMX1(255, 255, 255, 255)
+
+23
+00:00:26,428 --> 00:00:27,085
+HUE1(0,255,0,0);HUE2(0,255,0,0);HUE3(0,255,0,0);DMX1(0, 0, 0, 0)
+
+24
+00:00:27,142 --> 00:00:27,942
+HUE1(5000,255,100,0);HUE2(5000,255,100,0);HUE3(5000,255,100,0);DMX1(255, 0, 0, 0)
+
+25
+00:00:28,085 --> 00:00:28,885
+HUE1(10000,255,100,0);HUE2(10000,255,100,0);HUE3(10000,255,100,0);DMX1(0, 255, 0, 0)
+
+26
+00:00:29,171 --> 00:00:30,114
+HUE1(15000,255,100,0);HUE2(15000,255,100,0);HUE3(15000,255,100,0);DMX1(0, 0, 255, 0)
+
+27
+00:00:30,114 --> 00:00:31,142
+HUE1(20000,255,100,0);HUE2(20000,255,100,0);HUE3(20000,255,100,0);DMX1(0, 0, 0, 255)
+
+28
+00:00:31,142 --> 00:00:32,228
+HUE1(25000,255,100,0);HUE2(25000,255,100,0);HUE3(25000,255,100,0);DMX1(255, 0, 255, 0)
+
+29
+00:00:32,342 --> 00:00:34,342
+HUE1(30000,255,100,0);HUE2(30000,255,100,0);HUE3(30000,255,100,0);DMX1(255, 255, 255, 255)
+
+30
+00:00:34,485 --> 00:00:37,400
+HUE1(0,255,0,0);HUE2(0,255,0,0);HUE3(0,255,0,0);DMX1(0, 0, 0, 0)
+
