--- tracks.json
@@ -1,8 +1,0 @@
-[
-    {"Path":"01_TalesOfBath_LOOP.mlp","Name":"01_TalesOfBath_LOOP.mlp","Size":482697618,"MimeType":"application/octet-stream","ModTime":"2018-11-05T14:30:33.088Z","IsDir":false,"ID":"1bQVTUq-HteUXoxbVi5VlQYtkFT3qGiSL"},
-    {"Path":"02_Planets_Part1_Treatment.mlp","Name":"02_Planets_Part1_Treatment.mlp","Size":1606167210,"MimeType":"application/octet-stream","ModTime":"2018-11-05T14:04:12.003Z","IsDir":false,"ID":"1NTfNqgRzFdFwp0LzQIWKcm8ib3SBQNq-"},
-    {"Path":"02_TalesOfBath_POEM.mlp","Name":"02_TalesOfBath_POEM.mlp","Size":719234778,"MimeType":"application/octet-stream","ModTime":"2018-11-05T14:30:47.826Z","IsDir":false,"ID":"173nzYzMNTf8GLgVSng29DcSukdLG4_YK"},
-    {"Path":"03_Planets_Loop.mlp","Name":"03_Planets_Loop.mlp","Size":980617654,"MimeType":"application/octet-stream","ModTime":"2018-11-05T14:31:06.602Z","IsDir":false,"ID":"1KuyXneu_xRGI41qweTFGSCRy_X9bPbrn"},
-    {"Path":"03_TalesOfBath_STORY.mlp","Name":"03_TalesOfBath_STORY.mlp","Size":1060826876,"MimeType":"application/octet-stream","ModTime":"2018-11-05T14:32:20.545Z","IsDir":false,"ID":"1JgdJ2DGIVCkEAp7nGf9fUYNEN_aw9We5"},
-    {"Path":"5.1_AAC_Test.mp4","Name":"5.1_AAC_Test.mp4","Size":6134724,"MimeType":"video/mp4","ModTime":"2018-11-05T15:04:27.242Z","IsDir":false,"ID":"1K5yOq4Xsgen2Q6NGIgBncRnikslNlWbV"}
-    ]
